docs: bring README, AGENTS, docs/, and specs back in sync with v0.15.2
The audit caught a long backlog of doc/spec drift accumulated through
v0.15 → v0.15.2. This is the cleanup PR.

Summary of fixes:

README.md / AGENTS.md / docs/src/agents.md
- "Five plugins" → "three" (v0.15.2 dropped templates-remote and doctor)
- `templates-search` (hyphenated) → `templates search` (subcommand)
- Doctor JSON shape gains `informational: bool` per Section
- Drop the `templates-search` / `doctor-tools` rows from the
  --json-surface tables; add the new `templates search --json` row
- `fledge lane …` (singular) → `fledge lanes …` (canonical) throughout

docs/src/cli-reference.md
- Plugin-commands header updated (5 entries, not 8)
- `templates-search` / `templates-publish` plugin sections rewritten
  as proper `templates search` / `templates publish` subcommand docs
- `doctor` section rewritten to describe the four sections, the
  informational Toolchains addition, and the new JSON shape
- `doctor-tools` section removed (command no longer exists)
- `--defaults` description and the default-plugins table reflect
  the v0.15.2 three-plugin set

docs/src/doctor.md
- Full rewrite. The "Toolchain (project-type-specific)" + Dependencies
  framing is gone — replaced with the actual four sections (fledge,
  Git, AI, Toolchains) and a clear note that Toolchains is
  informational. New JSON example. New "what's probed" table.

docs/src/pillars.md
- Six-pillars description updated for templates' six subcommands
- Default-plugins list trimmed to three; explanatory note about the
  two re-absorbed plugins

docs/src/templates.md
- "Search GitHub (plugin)" rewritten as a core subcommand
- "Publish (plugin)" rewritten as `templates publish` (core)
- Drop the `gh repo create --push` description (we use publish.rs
  helpers directly, not gh CLI)

docs/src/getting-started/{quick-start,existing-projects}.md
- Annotate which commands are core vs plugin so a fresh-install user
  doesn't get confused when `fledge metrics` errors with "unknown
  command"

docs/src/github-integration.md
- `--ai-with-model` → `--with-model` (the actual flag)

docs/src/use-cases.md
- "+ 4 others" → "+ 2 others"
- `fledge lane ci` → `fledge lanes run ci`

docs/src/review.md
- Mention the v0.15.2 metrics rewrite (tokei now linked as a library;
  no separate `cargo install tokei` needed)

docs/src/troubleshooting.md
- GitHub-token paragraph annotates which commands are core vs plugin

specs/doctor/{context,requirements,tasks,testing}.md
- Full pass to reflect doctor v6: project-type detection is gone, the
  Toolchains informational section is in, all 16 probes enumerated,
  10s timeout invariant documented

specs/search/{context,requirements,tasks,testing}.md
- Reframed as "this is a library" — describes the helpers, not a
  user-facing command. Removes references to `run`/`SearchOptions`/
  `search_github_ex` which were stripped in v0.15. Manual-testing
  examples updated to `fledge templates search`/`lanes search`/
  `plugins search`

specs/publish/{context,requirements,testing}.md
- Same library-not-subcommand framing. Removes `validate_*` /
  `run_rejects_no_token` test entries that no longer correspond
  to anything in the source

specs/templates/{context,requirements}.md
- "2 built-in templates" → 8 (rust-cli, ts-bun, python-cli, go-cli,
  ts-node, static-site, kotlin-kmp, kotlin-ktor-api)
- Out-of-scope `create-template` and `publish` notes updated to
  point at the sibling `templates create` / `templates publish`
  subcommands

specs/main/main.spec.md
- `fledge --version` example: `0.8.0` → `0.15.2`

specs/spinner/{spinner.spec,testing}.md
- Behavioral examples updated away from `fledge checks` /
  `fledge prs` (plugin commands now) to `fledge templates search`
  and `fledge work pr` (core)

specs/plugin/{plugin.spec,requirements}.md
- Fix `pluginss` typos (×6)
- Fix `fledge plugin list` → `fledge plugins list` (canonical)
- Fix duplicate-numbered invariant 9 (renumber 10–14)

specs/work/context.md
- Note that `fledge checks --json` / `fledge prs --json` come from
  the `fledge-plugin-github` plugin, not core

Tests reorganization
- tests/integration.rs (2749 LOC monolith with 157 tests across
  13 implicit sections) split into per-module files matching the
  spec layout: tests/{templates,config,run,lanes,doctor,validate,
  spec,changelog,main}.rs. Shared helpers moved to
  tests/common/mod.rs. Each test file uses `mod common; use
  common::*;` to pull them in.
- tests/fledge.test.ts and tests/placeholder.test.ts deleted —
  fledge has no bun integration; the dummy files served no purpose.
- All 157 integration tests pass under their new file homes.

Validation
- `cargo test`: 663 unit + 7 + 5 + 2 + 7 + 69 + 17 + 27 + 17 + 6 = 820
  passing (1 ignored)
- `cargo clippy --tests -- -D warnings`: clean
- `cargo fmt --check`: clean
- `cargo run -- spec check --strict`: 29 specs, 0 errors, 0 warnings
