+++ docs/src/getting-started/existing-projects.md
@@ -58,20 +58,23 @@ Once `fledge.toml` exists, it takes full precedence over auto-detection.
 
 These commands work in any git repo regardless of how the project was created:
 
-| Command | What it does |
-|---------|-------------|
-| `fledge run` | Task runner (zero-config or from fledge.toml) |
-| `fledge lanes` | Workflow pipelines |
-| `fledge review` | AI code review of your current branch |
-| `fledge ask` | Ask questions about your codebase |
-| `fledge work` | Feature branch and PR workflow |
-| `fledge checks` | CI/CD status |
-| `fledge changelog` | Changelog from git tags |
-| `fledge issues` | GitHub issues |
-| `fledge prs` | Pull requests |
-| `fledge metrics` | Code stats (LOC, churn, test ratio) |
-| `fledge deps` | Dependency health |
-| `fledge doctor` | Environment diagnostics |
+| Command | What it does | Where it lives |
+|---------|-------------|----------------|
+| `fledge run` | Task runner (zero-config or from fledge.toml) | core |
+| `fledge lanes` | Workflow pipelines | core |
+| `fledge review` | AI code review of your current branch | core |
+| `fledge ask` | Ask questions about your codebase | core |
+| `fledge work` | Feature branch and PR workflow | core |
+| `fledge changelog` | Changelog from git tags | core |
+| `fledge doctor` | Environment diagnostics (incl. Toolchains) | core |
+| `fledge templates search`/`publish` | Browse / push community templates | core |
+| `fledge checks` | CI/CD status | `fledge-plugin-github` |
+| `fledge issues` | GitHub issues | `fledge-plugin-github` |
+| `fledge prs` | Pull requests | `fledge-plugin-github` |
+| `fledge metrics` | Code stats (LOC, churn, test ratio) | `fledge-plugin-metrics` |
+| `fledge deps` | Dependency health | `fledge-plugin-deps` |
+
+Plugin commands require `fledge plugins install --defaults` once.
 
 ## Turn Your Project into a Template
 
@@ -98,5 +101,5 @@ fledge templates publish ./my-stack
 1. **Start using fledge today**: `cd your-project && fledge run test`
 2. **Optionally lock in config**: `fledge run --init` to generate `fledge.toml`
 3. **Set up lanes**: `fledge lanes init` for CI pipelines
-4. **Use the full toolkit**: `fledge review`, `fledge work start feature-x`, `fledge checks`
+4. **Use the full toolkit**: `fledge review`, `fledge work start feature-x`, `fledge checks` (plugin)
 5. **Create templates**: Once you have a setup you like, `fledge templates create` to reuse it
