+++ specs/doctor/tasks.md
@@ -5,9 +5,9 @@ spec: doctor.spec.md
 ## Tasks
 
 - [x] Write spec files for doctor module
-- [x] Implement environment diagnostics (toolchain, deps, git, AI)
-- [x] Add per-language toolchain checks
-- [x] Add actionable fix suggestions for failing checks
-- [x] Add JSON output mode
+- [x] Implement environment diagnostics (fledge config, git, AI)
+- [x] Add JSON output mode with informational-section flag
 - [x] Wire up CLI subcommand in main.rs
+- [x] (v0.15) Strip per-project-type toolchain probes
+- [x] (v0.15.2) Re-absorb toolchain probes from `fledge-plugin-doctor` as the informational `Toolchains` section
 - [x] Run verification suite
