+++ specs/search/testing.md
@@ -6,36 +6,31 @@
 |------|-------------|
 | `parse_search_response_valid` | Parses a valid GitHub search API JSON response into `Vec<SearchResult>` |
 | `parse_search_response_empty` | Handles empty `items` array gracefully |
-| `parse_search_response_missing_fields` | Handles repos with null description |
+| `parse_search_response_missing_fields` | Defaults missing description / stars / topics to sensible values |
+| `parse_search_response_skips_no_owner` | Items without `owner.login` are silently skipped |
 | `build_search_query_no_keyword` | Builds correct query with only topic filter |
 | `build_search_query_with_keyword` | Builds correct query combining topic and keyword |
-| `format_stars_abbreviation` | Formats star counts (e.g. 1500 -> "1.5k") |
-| `search_result_usage_hint` | Generates correct `fledge init -t owner/repo` hint |
-| `json_output_format` | Serializes results to expected JSON structure |
+| `build_search_query_with_author` | Adds `user:` filter when an author is provided |
+| `format_stars_abbreviation` | Formats star counts (e.g. 1500 → "1.5k", 123456 → "123k") |
+| `urlencod_unreserved` | Keeps `[A-Za-z0-9-_.~]` unencoded per RFC 3986 |
+| `urlencod_special` | Percent-encodes spaces as `%20` (not `+`), colons as `%3A` |
 
 ## Integration Tests
 
-Integration tests require network access and are gated behind `#[ignore]` — run with `cargo test -- --ignored`.
-
-| Test | Description |
-|------|-------------|
-| `live_search_returns_results` | Performs actual GitHub API search (ignored by default) |
+The user-facing surfaces (`fledge templates search`, `fledge lanes search`, `fledge plugins search`) drive these helpers and are tested at the CLI level in `tests/templates.rs`, `tests/lanes.rs`, etc. Live network calls are gated behind `#[ignore]` — run with `cargo test -- --ignored`.
 
 ## Manual Testing
 
 ```bash
-# Basic search
-fledge search
+# Templates flavor (the user-facing search wired to these helpers)
+fledge templates search                    # browse all fledge-template repos
+fledge templates search rust               # filter by keyword
+fledge templates search --author CorvidLabs
+fledge templates search --limit 5 --json
 
-# Search with keyword
-fledge search rust
+# Lanes flavor
+fledge lanes search
 
-# JSON output
-fledge search --json
-
-# With limit
-fledge search --limit 5
-
-# Combined
-fledge search python --limit 3 --json
+# Plugins flavor
+fledge plugins search deploy
 ```
