+++ specs/work/context.md
@@ -19,6 +19,6 @@ Fledge is evolving from scaffolding to a project lifecycle tool. `fledge work` p
 - `--issue` flag prepends the issue number to the name portion for GitHub issue linking
 - Sanitize branch names (lowercase, replace non-alphanumeric with hyphens) to avoid git issues
 - `generate_title_from_branch` dynamically strips any known type prefix rather than hardcoding a few
-- **v6: `--json` added to all three subcommands** so agents can drive the workflow without parsing human-formatted output. Rationale: agents using `fledge work start` need the branch name back; agents using `fledge work pr` need the PR URL and number; `fledge work status` should give a shape compatible with `fledge checks --json` and `fledge prs --json`. Each JSON payload is a single top-level object.
+- **v6: `--json` added to all three subcommands** so agents can drive the workflow without parsing human-formatted output. Rationale: agents using `fledge work start` need the branch name back; agents using `fledge work pr` need the PR URL and number; `fledge work status` should give a shape compatible with the GitHub-plugin commands `fledge checks --json` and `fledge prs --json` (provided by `fledge-plugin-github`). Each JSON payload is a single top-level object.
 - In JSON mode, spinners and ✅ output are suppressed entirely so the only stdout content is the JSON payload. Errors still go to stderr, and the process still exits non-zero on failure — so `if fledge work pr --json; then ... fi` works as expected.
 - `status` now also reports `behind` (commits the base has that the branch doesn't). Cheap and agent-useful.
